--- a/ring/example/local/go.mod
+++ b/ring/example/local/go.mod
@@ -1,8 +1,6 @@
 module github.com/grafana/dskit/ring/example/local
 
-go 1.21
-
-toolchain go1.22.4
+go 1.22
 
 require (
 	github.com/go-kit/log v0.2.1