@56quarters 2025-02-27
```
$ benchstat old.txt new.txt 
goos: linux
goarch: amd64
pkg: github.com/grafana/dskit/cache
cpu: Intel(R) Core(TM) i9-10885H CPU @ 2.40GHz
                                        │   old.txt    │               new.txt               │
                                        │    sec/op    │   sec/op     vs base                │
MemcachedJumpHashSelector_PickServer-16   162.75n ± 1%   31.65n ± 1%  -80.55% (p=0.000 n=20)

                                        │  old.txt   │              new.txt               │
                                        │    B/op    │   B/op     vs base                 │
MemcachedJumpHashSelector_PickServer-16   24.00 ± 0%   0.00 ± 0%  -100.00% (p=0.000 n=20)

                                        │  old.txt   │               new.txt               │
                                        │ allocs/op  │ allocs/op   vs base                 │
MemcachedJumpHashSelector_PickServer-16   1.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=20)
```